+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Use a more precise timestamp when generating temporary directory names to
 permit more than one concurrent checkpointing attempt per second.
+- Automatic import of nest_asyncio.
 
 ## [0.1.4] - 2022-03-15